fix manual conversion functions
