+++ pkg/apis/dashboard/v2alpha1/types.go
@@ -2,7 +2,6 @@ package v2alpha1
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	conversion "k8s.io/apimachinery/pkg/conversion"
 
 	data "github.com/grafana/grafana-plugin-sdk-go/experimental/apis/data/v0alpha1"
 	common "github.com/grafana/grafana/pkg/apimachinery/apis/common/v0alpha1"
@@ -164,23 +163,4 @@ type AnnotationActions struct {
 	CanAdd    bool `json:"canAdd"`
 	CanEdit   bool `json:"canEdit"`
 	CanDelete bool `json:"canDelete"`
-}
-
-func Convert_v0alpha1_Unstructured_To_v2alpha1_DashboardSpec(in *common.Unstructured, out *DashboardSpec, s conversion.Scope) error {
-	out.Unstructured = *in
-
-	t := in.Object["title"]
-	title, ok := t.(string)
-	if !ok {
-		// do not force the title field
-		return nil
-	}
-	out.Title = title
-
-	return nil
-}
-
-func Convert_v2alpha1_DashboardSpec_To_v0alpha1_Unstructured(in *DashboardSpec, out *common.Unstructured, s conversion.Scope) error {
-	out = &in.Unstructured
-	return nil
 }
